Bugfix for test_dispatch: checks if dump DB file exists before trying to delete it

+++ tests/test_dispatch.py
@@ -88,7 +88,8 @@ class TestSQLiteDispatchDelete(TestSQLiteBase):
 
     def tearDown(self):
         super(TestSQLiteDispatchDelete, self).tearDown()
-        os.unlink(self.test_dump_filename)
+        if os.path.exists(self.test_dump_filename):
+            os.unlink(self.test_dump_filename)
 
     def test_delete(self):
         test_task = Task()
